feature: No circular import anymore, but some tests fail.
