+++ tests/test_ensembles.py
@@ -2,8 +2,9 @@ from random import randint
 
 import numpy as np
 
+from fold.composites.columns import PerColumnEnsemble
 from fold.composites.concat import Pipeline
-from fold.composites.ensemble import Ensemble, PerColumnEnsemble
+from fold.composites.ensemble import Ensemble
 from fold.loop import backtest, train
 from fold.models.dummy import DummyClassifier
 from fold.splitters import ExpandingWindowSplitter
