+++ tests/test_transformations.py
@@ -1,11 +1,12 @@
 import numpy as np
 
+from fold.composites.columns import PerColumnTransform
+from fold.composites.concat import TransformColumn
 from fold.loop import backtest, train
 from fold.loop.types import TrainMethod
 from fold.models.base import Model
 from fold.splitters import ExpandingWindowSplitter
 from fold.transformations import Identity, SelectColumns
-from fold.transformations.columns import PerColumnTransform, TransformColumn
 from fold.transformations.difference import Difference
 from fold.transformations.lags import AddLagsY
 from fold.utils.tests import generate_sine_wave_data
